pyproject.toml
[project]
name = "evolve-o-mart"
version = "1.0.0"
description = "EVOLVE-O-MART - An MCP-powered evolving product store"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp==3.0.0b1",
    "pydantic>=2.0",
    "click>=8.0",
    "py-key-value-aio>=0.3.0",
    "anthropic>=0.40",
    "google-genai>=1.60.0",
    "elasticsearch>=9.2.1",
    "aiohttp>=3.13.3",
]

[project.scripts]
evolve-o-mart = "evolve_o_mart.server:cli"

[build-system]
requires = ["uv_build>=0.7.5"]
build-backend = "uv_build"

[tool.uv]
package = true

[dependency-groups]
dev = [
    "basedpyright>=1.37.1",
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.14.13",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 140
target-version = "py312"
extend-exclude = [
    ".git",
    ".mypy_cache",
    ".pytest_cache",
    ".ruff_cache",
    ".venv",
    "venv",
    "__pycache__",
    "build",
    "dist",
    "reference",
    "**/*.old.py",
]

[tool.ruff.lint]
extend-select = [
    "A", "ARG", "B", "C4", "COM", "DTZ", "E", "EM", "F", "FURB", "I",
    "LOG", "N", "PERF", "PIE", "PLR", "PLW", "PT", "PTH", "Q", "RET",
    "RSE", "RUF", "S", "SIM", "TC", "TID", "TRY", "UP", "W",
    "D",
    "ANN",
    "PLE",
    "T10",
    "ISC",
    "FA",
    "PGH",
    "INP",
]
ignore = [
    "COM812",
    "D203",
    "D213",
    "D413",
    "D100",
    "A002",
    "ANN401",
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "S101",
    "PLR2004",
    "INP001",
]
"src/evolve_o_mart/gemini/sampling.py" = [
    "TRY003", "EM101", "D107", "D102", "ARG002", "PERF401",
]
"src/evolve_o_mart/storage/elasticsearch.py" = [
    "D103",
]

[tool.basedpyright]
pythonVersion = "3.12"
typeCheckingMode = "recommended"
reportImportCycles = "error"
reportMissingTypeStubs = false
reportExplicitAny = false
